feat(baremetal): enable RV32 FPU and migrate emitrtl to playground
- rv32imc → rv32imfc; start.S uses RVTEST_RV32UF macro
- vecAdd arrays: uint32_t → float
- Remove local emitrtl submodule; point build.sc to playground copy
- Fix exit_syscall.h inline asm to use named register constraints
- Update sim exe path to include chisel_gen_rtl/ intermediate dir
- Qualify TOP_LAZY_MODULE_CONFIG names with config suffix

+++ src/main/resources/baremetal/vecAdd/bare_main.c
@@ -2,9 +2,9 @@
 #include <stdint.h>
 #include "exit_syscall.h"
 
-uint32_t dataA[] = {0,1,2,3,4,5,6,7};
-uint32_t dataB[] = {0,1,2,3,4,5,6,7};
-uint32_t dataC[] = {0,2,4,6,8,10,12,14};
+float dataA[] = {0,1,2,3,4,5,6,7};
+float dataB[] = {0,1,2,3,4,5,6,7};
+float dataC[] = {0,2,4,6,8,10,12,14};
 
 void bare_main(void)
 {
